+++ CHANGELOG.md
@@ -19,13 +19,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - :interface- special attribute support in tags
 - async_result/1
 - render_upload support in LiveViewNativeTest
-- suppport single quotes to wrap attribute values in template parser
+- support single quotes to wrap attribute values in template parser
 - LVN Commands
+- LiveViewNative.Template.Engine
 
 ### Changed
 
 - `LiveViewNative.Component` no longer imports `Phoenix.Component.to_form/2`
 - `LiveViewNative.LiveView` now requires the `dispatch_to` function to determine which module will be used for rendering
+- Migrated many functions out of LiveViewNative.TagEngine to LiveViewNative.Template.Engine
 
 ### Fixed
