+++ test/support/clients/gameboy.ex
@@ -3,6 +3,6 @@ defmodule LiveViewNativeTest.GameBoy do
     format: :gameboy,
     component: LiveViewNativeTest.GameBoy.Component,
     module_suffix: :GameBoy,
-    template_engine: LiveViewNative.Engine,
+    template_engine: LiveViewNative.Template.Engine,
     test_client: %LiveViewNativeTest.GameBoy.TestClient{}
 end
